+++ Sasquatch/SasquatchUITests/AppCenterUITests.swift
@@ -47,9 +47,6 @@ class AppCenterUITests: XCTestCase {
     // Push.
     XCTAssertFalse(app.tables["App Center"].switches["Push Enabled"].boolValue)
     
-    // Auth.
-    XCTAssertFalse(app.tables["App Center"].switches["Auth Enabled"].boolValue)
-    
     // Analytics.
     app.tabBars.buttons["Analytics"].tap()
     XCTAssertFalse(app.tables["Analytics"].switches["Set Enabled"].boolValue)
@@ -69,9 +66,6 @@ class AppCenterUITests: XCTestCase {
     // All services should be enabled.
     // Push.
     XCTAssertTrue(app.tables["App Center"].switches["Push Enabled"].boolValue)
-    
-    // Auth.
-    XCTAssertTrue(app.tables["App Center"].switches["Auth Enabled"].boolValue)
     
     // Analytics.
     app.tabBars.buttons["Analytics"].tap()